add HeartHub/SignUpViewController.swift
